Labyrinth objects are now positioned using floats.

+++ LabyrinthLib/L/LabyrinthObject.cs
@@ -9,16 +9,16 @@ namespace LabyrinthLib.L
 {
     public abstract class LabyrinthObject
     {
-        protected int x;
-        protected int y;
-        protected int w;
-        protected int h;
-        public int X { get { return x; } set { x = value; } }
-        public int Y { get { return y; } set { y = value; } }
-        public int W { get { return w; } set { w = value; } }
-        public int H { get { return h; } set { h = value; } }
+        protected float x;
+        protected float y;
+        protected float w;
+        protected float h;
+        public float X { get { return x; } set { x = value; } }
+        public float Y { get { return y; } set { y = value; } }
+        public float W { get { return w; } set { w = value; } }
+        public float H { get { return h; } set { h = value; } }
 
-        public LabyrinthObject(int x, int y, int w, int h)
+        public LabyrinthObject(float x, float y, float w, float h)
         {
             this.x = x;
             this.y = y;
